+++ src/main/java/com/enterprise/livro/entities/Editora.java
@@ -10,23 +10,23 @@ public class Editora {
 	
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Id
-	private Integer id;
+	private Long id;
 	private String nome;
 	
 	public Editora() {
 	}
 
-	public Editora(Integer id, String nome) {
+	public Editora(Long id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
